feat: 로그인을 하지 않은 사용자가 로그아웃을 시도할 때 에러를 던지는 기능 추가

+++ src/main/java/balancetalk/module/member/application/MemberService.java
@@ -111,9 +111,10 @@ public class MemberService {
         Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
         if (principal instanceof UserDetails) {
             String username = ((UserDetails) principal).getUsername();
-            log.info("Before logout={}" , redisService.getValues(username));
+            if (redisService.getValues(username) == null) {
+                throw new BalanceTalkException(ErrorCode.UNAUTHORIZED_LOGOUT);
+            }
             redisService.deleteValues(username);
-            log.info("After logout={}" , redisService.getValues(username));
         }
     }
